(tests) Rename Perlang.Tests to Perlang.Tests.Integration
Also added a new `Perlang.Tests` project, with a more unit-test-oriented
approach. The `Perlang.Tests` project contains a single test for now,
mostly to just illustrate how testing can be done here, and also help
ensure unit-testability of our code.
