+++ Perlang.sln
@@ -6,11 +6,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.Common", "Perlang.C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.Parser", "Perlang.Parser\Perlang.Parser.csproj", "{F1F9229C-D764-4A38-BEA2-1F86A1C2FF1A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.Tests", "Perlang.Tests\Perlang.Tests.csproj", "{1874E267-5D6F-4EA4-8938-F2A0D19F27B1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.Tests.Integration", "Perlang.Tests.Integration\Perlang.Tests.Integration.csproj", "{1874E267-5D6F-4EA4-8938-F2A0D19F27B1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.ConsoleApp", "Perlang.ConsoleApp\Perlang.ConsoleApp.csproj", "{EC950B39-0187-41BD-A9AA-FE18F73BD30F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.Stdlib", "Perlang.Stdlib\Perlang.Stdlib.csproj", "{20267530-E991-4936-BFDC-0C5A1EDE90E7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Perlang.Tests", "Perlang.Tests\Perlang.Tests.csproj", "{CD78DA11-CB31-4892-8F96-3B7FD7015B83}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -42,5 +44,9 @@ Global
 		{20267530-E991-4936-BFDC-0C5A1EDE90E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{20267530-E991-4936-BFDC-0C5A1EDE90E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{20267530-E991-4936-BFDC-0C5A1EDE90E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CD78DA11-CB31-4892-8F96-3B7FD7015B83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD78DA11-CB31-4892-8F96-3B7FD7015B83}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD78DA11-CB31-4892-8F96-3B7FD7015B83}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD78DA11-CB31-4892-8F96-3B7FD7015B83}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
